+++ Source/WebScheduler.Api.Core/Constants/Cluster.cs
@@ -1,0 +1,7 @@
+namespace WebScheduler.Api.Core.Constants;
+
+public static class Cluster
+{
+    public const string ClusterId = "ClusterId";
+    public const string ServiceId = "ServiceId";
+}
