+++ Source/WebScheduler.Api.Core/Mappers/ScheduledTaskMetaDataToScheduledTaskMapper.cs
@@ -1,0 +1,44 @@
+namespace WebScheduler.Api.Core.Mappers;
+
+using Boxed.Mapping;
+using WebScheduler.Abstractions.Grains.Scheduler;
+using WebScheduler.Api.Core.Models;
+
+public class ScheduledTaskMetaDataToScheduledTaskMapper : IMapper<GuidIdWrapper<ScheduledTaskMetadata>, ScheduledTask>, IMapper<ScheduledTask, GuidIdWrapper<ScheduledTaskMetadata>>
+{
+    // TODO: this is broken
+    public void Map(GuidIdWrapper<ScheduledTaskMetadata> source, ScheduledTask destination)
+    {
+        ArgumentNullException.ThrowIfNull(source);
+        ArgumentNullException.ThrowIfNull(destination);
+
+        destination.ScheduledTaskId = source.Id;
+        destination.IsEnabled = source.Value.IsEnabled;
+        destination.Description = source.Value.Description;
+        destination.Name = source.Value.Name;
+        destination.CreatedAt = source.Value.CreatedAt;
+        destination.ModifiedAt = source.Value.ModifiedAt;
+        destination.CronExpression = source.Value.CronExpression;
+        destination.NextRunAt = source.Value.NextRunAt;
+        destination.LastRunAt = source.Value.LastRunAt;
+        destination.HttpTriggerProperties = source.Value.HttpTriggerProperties;
+        destination.TriggerType = source.Value.TriggerType;
+    }
+
+    public void Map(ScheduledTask source, GuidIdWrapper<ScheduledTaskMetadata> destination)
+    {
+        ArgumentNullException.ThrowIfNull(source);
+        ArgumentNullException.ThrowIfNull(destination);
+        //var 
+        //destination.Id = source.ScheduledTaskId;
+        //destination.IsEnabled = source.IsEnabled;
+        //destination.Description = source.Value.Description;
+        //destination.Name = source.Value.Name;
+        //destination.Url = new Uri(this.linkGenerator.GetUriByRouteValues(
+        //    this.httpContextAccessor.HttpContext!,
+        //    ScheduledTasksControllerRoute.GetScheduledTask,
+        //    new { source.Id })!);
+    }
+}
+
+public record GuidIdWrapper<TValue>(Guid Id, TValue Value);
