+++ Source/WebScheduler.Api.Core/Models/ScheduledTask.cs
@@ -1,17 +1,18 @@
-namespace WebScheduler.Api.Models.ViewModels;
+namespace WebScheduler.Api.Core.Models;
 
+using Cronos;
 using WebScheduler.Abstractions.Grains.Scheduler;
 
 /// <summary>
-/// A make and model of car.
+/// A scheduled task.
 /// </summary>
 public class ScheduledTask
 {
-    /// <summary>
-    /// Gets or sets the scheduled task's unique identifier.
-    /// </summary>
-    /// <example>1AE8039D-E152-46E6-ADD6-BA14F1C5EE4A</example>
     public Guid ScheduledTaskId { get; set; }
+
+    public DateTime CreatedAt { get; set; }
+
+    public DateTime ModifiedAt { get; set; }
 
     /// <summary>
     /// The name of the scheduled task.
@@ -28,13 +29,18 @@ public class ScheduledTask
     /// </summary>
     public bool IsEnabled { get; set; }
 
+    /// <summary>
+    /// When the task was last run.
+    /// </summary>
     public DateTime? LastRunAt { get; set; }
-    public DateTime? NextRunAt { get; set; }
-    public DateTime ModifiedAt { get; set; }
-    public DateTime CreatedAt { get; set; }
 
     /// <summary>
-    /// The cron expression to use for the task schedule.
+    /// When the task will next run.
+    /// </summary>
+    public DateTime? NextRunAt { get; set; }
+
+    /// <summary>
+    /// The cron expression to use for the task schedule. We only want users to input expressions with per-minute granularity, hence the <seealso cref="CronFormat.Standard"/>.
     /// </summary>
     public string CronExpression { get; set; } = string.Empty;
 
@@ -47,10 +53,4 @@ public class ScheduledTask
     /// The properties required to support a <see cref="TriggerType"/> value of <seealso cref="TaskTriggerType.HttpTrigger"/>
     /// </summary>
     public HttpTriggerProperties HttpTriggerProperties { get; set; } = new();
-
-    /// <summary>
-    /// Creates a shallow copy of the object instance.
-    /// </summary>
-    /// <returns>The copy.</returns>
-    public ScheduledTask ShallowCopy() => (ScheduledTask)this.MemberwiseClone();
 }
