+++ Source/WebScheduler.Api.Core/Options/StorageOptions.cs
@@ -1,0 +1,7 @@
+namespace WebScheduler.Api.Core.Options;
+
+public class StorageOptions
+{
+    public string Invariant { get; set; } = default!;
+    public string ConnectionString { get; set; } = default!;
+}
