+++ Source/WebScheduler.Api.Core/Repositories/EnumerableExtensions.cs
@@ -1,0 +1,20 @@
+namespace WebScheduler.Api.Core.Repositories;
+
+public static class EnumerableExtensions
+{
+    public static IEnumerable<T> If<T>(
+        this IEnumerable<T> enumerable,
+        bool condition,
+        Func<IEnumerable<T>, IEnumerable<T>> action)
+    {
+        ArgumentNullException.ThrowIfNull(enumerable);
+        ArgumentNullException.ThrowIfNull(action);
+
+        if (condition)
+        {
+            return action(enumerable);
+        }
+
+        return enumerable;
+    }
+}
