+++ Source/WebScheduler.Api.Core/WebScheduler.Client.Core.csproj
@@ -1,0 +1,55 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup Label="Build">
+    <TargetFramework>net6.0</TargetFramework>
+    <GenerateDocumentationFile>True</GenerateDocumentationFile>
+    <IsPackable>true</IsPackable>
+  </PropertyGroup>
+  <PropertyGroup Label="Package">
+    <Authors>Elan Hasson</Authors>
+    <Company>Elan Hasson</Company>
+    <Product>Web Scheduler Api Core</Product>
+    <Description>The API Core for the Web Scheduler</Description>
+    <Copyright>Copyright © Elan Hasson. All rights Reserved</Copyright>
+  </PropertyGroup>
+  <PropertyGroup>
+    <!-- Workaround to build this project: https://github.com/dotnet/templating/issues/1438 -->
+    <GitHub>true</GitHub>
+    <IsTrimmable>false</IsTrimmable>
+    <Sign>false</Sign>
+    <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
+    <RepositoryUrl>https://github.com/web-scheduler/web-scheduler</RepositoryUrl>
+    <RepositoryType>git</RepositoryType>
+    <PackageProjectUrl>https://github.com/web-scheduler/web-scheduler</PackageProjectUrl>
+    <PackageReleaseNotes>https://github.com/web-scheduler/web-scheduler/releases</PackageReleaseNotes>
+
+    <PackageLicenseExpression>MIT</PackageLicenseExpression>
+    <PackageRequireLicenseAcceptance>true</PackageRequireLicenseAcceptance>
+    <PackageTags>orleans;cron;scheduler</PackageTags>
+
+  </PropertyGroup>
+  <ItemGroup Label="Package References">
+    <PackageReference Include="Boxed.Mapping" Version="6.0.0+build.458" />
+    <PackageReference Include="Dapper" Version="2.0.123" />
+    <PackageReference Include="Microsoft.Orleans.Client" Version="3.6.3" />
+    <PackageReference Include="Microsoft.Orleans.Clustering.AdoNet" Version="3.6.3" />
+    <PackageReference Include="Microsoft.Orleans.CodeGenerator.MSBuild" PrivateAssets="all" Version="3.6.3">
+      <IncludeAssets>runtime; build; native; contentfiles; analyzers</IncludeAssets>
+    </PackageReference>
+    <PackageReference Include="Microsoft.Orleans.Connections.Security" Version="3.6.3" />
+    <PackageReference Include="MySql.Data" Version="8.0.30" />
+    <PackageReference Update="Microsoft.VisualStudio.Threading.Analyzers" Version="17.2.32" />
+    <PackageReference Update="MinVer" Version="4.1.0" />
+    <PackageReference Include="WebScheduler.Abstractions" Version="0.1.37" />
+  </ItemGroup>
+
+  <ItemGroup Label="Project References">
+    <ProjectReference Include="..\WebScheduler.Api.Models\WebScheduler.Api.Models.csproj" />
+  </ItemGroup>
+
+  <ItemGroup Label="Project References For Debug">
+    <!--<ProjectReference Include="..\..\..\web-scheduler\Source\WebScheduler.Abstractions\WebScheduler.Abstractions.csproj" />-->
+  </ItemGroup>
+  
+  
+</Project>
