+++ Source/WebScheduler.Api.Http/GlobalSuppressions.cs
@@ -1,0 +1,13 @@
+// This file is used by Code Analysis to maintain SuppressMessage
+// attributes that are applied to this project.
+// Project-level suppressions either have no target or are given
+// a specific target and scoped to a namespace, type, member, etc.
+
+using System.Diagnostics.CodeAnalysis;
+
+[assembly: SuppressMessage("Performance", "CA1822:Mark members as static", Justification = "Method is an Controller action.", Scope = "member", Target = "~M:WebScheduler.Api.Http.Controllers.ScheduledTasksController.DeleteAsync(WebScheduler.Api.Http.Commands.ScheduledTask.DeleteScheduledTaskCommand,System.Guid,System.Threading.CancellationToken)~System.Threading.Tasks.Task{Microsoft.AspNetCore.Mvc.IActionResult}")]
+[assembly: SuppressMessage("Performance", "CA1822:Mark members as static", Justification = "Method is an Controller action.", Scope = "member", Target = "~M:WebScheduler.Api.Http.Controllers.ScheduledTasksController.GetAsync(WebScheduler.Api.Http.Commands.ScheduledTask.GetScheduledTaskCommand,System.Guid,System.Threading.CancellationToken)~System.Threading.Tasks.Task{Microsoft.AspNetCore.Mvc.IActionResult}")]
+[assembly: SuppressMessage("Performance", "CA1822:Mark members as static", Justification = "Method is an Controller action.", Scope = "member", Target = "~M:WebScheduler.Api.Http.Controllers.ScheduledTasksController.GetPageAsync(WebScheduler.Api.Http.Commands.ScheduledTask.GetScheduledTaskPageCommand,WebScheduler.Api.Models.ViewModels.PageOptions,System.Threading.CancellationToken)~System.Threading.Tasks.Task{Microsoft.AspNetCore.Mvc.IActionResult}")]
+[assembly: SuppressMessage("Performance", "CA1822:Mark members as static", Justification = "Method is an Controller action.", Scope = "member", Target = "~M:WebScheduler.Api.Http.Controllers.ScheduledTasksController.PostAsync(WebScheduler.Api.Commands.ScheduledTask.PostScheduledTaskCommand,WebScheduler.Api.Models.ViewModels.SaveScheduledTask,System.Threading.CancellationToken)~System.Threading.Tasks.Task{Microsoft.AspNetCore.Mvc.IActionResult}")]
+[assembly: SuppressMessage("Performance", "CA1822:Mark members as static", Justification = "Method is an Controller action.", Scope = "member", Target = "~M:WebScheduler.Api.Http.Controllers.ScheduledTasksController.PutAsync(WebScheduler.Api.Http.Commands.ScheduledTask.PutScheduledTaskCommand,System.Guid,WebScheduler.Api.Models.ViewModels.SaveScheduledTask,System.Threading.CancellationToken)~System.Threading.Tasks.Task{Microsoft.AspNetCore.Mvc.IActionResult}")]
+[assembly: SuppressMessage("Performance", "CA1822:Mark members as static", Justification = "Method is an Controller action.", Scope = "member", Target = "~M:WebScheduler.Api.Http.Controllers.ScheduledTasksController.PatchAsync(WebScheduler.Api.Http.Commands.ScheduledTask.PatchScheduledTaskCommand,System.Guid,Microsoft.AspNetCore.JsonPatch.JsonPatchDocument{WebScheduler.Api.Models.ViewModels.SaveScheduledTask},System.Threading.CancellationToken)~System.Threading.Tasks.Task{Microsoft.AspNetCore.Mvc.IActionResult}")]
