+++ Source/WebScheduler.Api.Models/WebScheduler.Api.Models.csproj
@@ -28,7 +28,7 @@
   <ItemGroup Label="Package References">
     <PackageReference Include="Cronos" Version="0.7.1" />
     <PackageReference Update="Microsoft.VisualStudio.Threading.Analyzers" Version="17.2.32" />
-    <PackageReference Include="WebScheduler.Abstractions" Version="0.1.36" />
+    <PackageReference Include="WebScheduler.Abstractions" Version="0.1.37" />
   </ItemGroup>
   <ItemGroup Label="Project References" Condition="'$(Configuration)' == 'Debug'">
     <!--<ProjectReference Include="..\..\..\web-scheduler\Source\WebScheduler.Abstractions\WebScheduler.Abstractions.csproj" />-->
