--- Source/WebScheduler.Api/Constants/Cluster.cs
@@ -1,7 +1,0 @@
-namespace WebScheduler.Api.Constants;
-
-public static class Cluster
-{
-    public const string ClusterId = "ClusterId";
-    public const string ServiceId = "ServiceId";
-}
