--- Source/WebScheduler.Api/Constants/ScheduledTasksControllerRoute.cs
@@ -1,15 +1,0 @@
-namespace WebScheduler.Api.Constants;
-
-public static class ScheduledTasksControllerRoute
-{
-    public const string DeleteScheduledTask = ControllerName.ScheduledTasks + nameof(DeleteScheduledTask);
-    public const string GetScheduledTask = ControllerName.ScheduledTasks + nameof(GetScheduledTask);
-    public const string GetScheduledTaskPage = ControllerName.ScheduledTasks + nameof(GetScheduledTaskPage);
-    public const string HeadScheduledTask = ControllerName.ScheduledTasks + nameof(HeadScheduledTask);
-    public const string HeadScheduledTaskPage = ControllerName.ScheduledTasks + nameof(HeadScheduledTaskPage);
-    public const string OptionsScheduledTask = ControllerName.ScheduledTasks + nameof(OptionsScheduledTask);
-    public const string OptionsScheduledTasks = ControllerName.ScheduledTasks + nameof(OptionsScheduledTasks);
-    public const string PatchScheduledTask = ControllerName.ScheduledTasks + nameof(PatchScheduledTask);
-    public const string PostScheduledTask = ControllerName.ScheduledTasks + nameof(PostScheduledTask);
-    public const string PutScheduledTask = ControllerName.ScheduledTasks + nameof(PutScheduledTask);
-}
