+++ Source/WebScheduler.Api/Options/ApplicationOptions.cs
@@ -2,8 +2,6 @@ namespace WebScheduler.Api.Options;
 
 using System.ComponentModel.DataAnnotations;
 using Microsoft.AspNetCore.Server.Kestrel.Core;
-using Orleans.Configuration;
-using WebScheduler.Server.Options;
 
 /// <summary>
 /// All options for the application.
@@ -29,10 +27,4 @@ public class ApplicationOptions
 
     [Required]
     public RedisOptions Redis { get; set; } = default!;
-
-    [Required]
-    public ClusterOptions Cluster { get; set; } = default!;
-
-    [Required]
-    public StorageOptions Storage { get; set; } = default!;
 }
