--- Source/WebScheduler.Api/Options/StorageOptions.cs
@@ -1,7 +1,0 @@
-namespace WebScheduler.Server.Options;
-
-public class StorageOptions
-{
-    public string Invariant { get; set; } = default!;
-    public string ConnectionString { get; set; } = default!;
-}
