--- Source/WebScheduler.Api/Repositories/IScheduledTaskRepository.cs
@@ -1,21 +1,0 @@
-namespace WebScheduler.Api.Repositories;
-
-using WebScheduler.Api.Models;
-
-public interface IScheduledTaskRepository
-{
-    Task<ScheduledTask> AddAsync(ScheduledTask scheduledTask, CancellationToken cancellationToken);
-
-    Task DeleteAsync(Guid scheduledTask, CancellationToken cancellationToken);
-
-    Task<ScheduledTask> GetAsync(Guid scheduledTaskId, CancellationToken cancellationToken);
-
-    Task<List<ScheduledTask>> GetScheduledTasksAsync(
-      int offset,
-      int pageSize,
-      CancellationToken cancellationToken);
-
-    Task<int> GetTotalCountAsync(CancellationToken cancellationToken);
-
-    Task<ScheduledTask> UpdateAsync(ScheduledTask scheduledTask, CancellationToken cancellationToken);
-}
