+++ WebScheduler.Api.sln
@@ -50,6 +50,10 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebScheduler.Api", "Source\
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebScheduler.Api.Models", "Source\WebScheduler.Api.Models\WebScheduler.Api.Models.csproj", "{A6B8CBAD-3284-4E4D-8444-32D968EB8A3B}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebScheduler.Client.Http", "Source\WebScheduler.Api.Http\WebScheduler.Client.Http.csproj", "{4BBA8EE4-B3AE-4591-9BF4-22C12E1D3DF9}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebScheduler.Client.Core", "Source\WebScheduler.Api.Core\WebScheduler.Client.Core.csproj", "{36E1CC43-47D6-4293-BE29-4536B92E4649}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -68,6 +72,14 @@ Global
 		{A6B8CBAD-3284-4E4D-8444-32D968EB8A3B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A6B8CBAD-3284-4E4D-8444-32D968EB8A3B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A6B8CBAD-3284-4E4D-8444-32D968EB8A3B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4BBA8EE4-B3AE-4591-9BF4-22C12E1D3DF9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4BBA8EE4-B3AE-4591-9BF4-22C12E1D3DF9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4BBA8EE4-B3AE-4591-9BF4-22C12E1D3DF9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4BBA8EE4-B3AE-4591-9BF4-22C12E1D3DF9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{36E1CC43-47D6-4293-BE29-4536B92E4649}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{36E1CC43-47D6-4293-BE29-4536B92E4649}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{36E1CC43-47D6-4293-BE29-4536B92E4649}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{36E1CC43-47D6-4293-BE29-4536B92E4649}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -80,6 +92,8 @@ Global
 		{5C84BC3D-BEFC-4CC2-85B8-C3C4E286368A} = {E5422D5C-ABD4-4150-BDF6-B823D746D157}
 		{9619BB00-2BA7-4834-AABE-5DB066C5F4A3} = {57BAE1BC-D418-4A95-B292-C0E9324B0A54}
 		{A6B8CBAD-3284-4E4D-8444-32D968EB8A3B} = {57BAE1BC-D418-4A95-B292-C0E9324B0A54}
+		{4BBA8EE4-B3AE-4591-9BF4-22C12E1D3DF9} = {57BAE1BC-D418-4A95-B292-C0E9324B0A54}
+		{36E1CC43-47D6-4293-BE29-4536B92E4649} = {57BAE1BC-D418-4A95-B292-C0E9324B0A54}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {4E464B40-8B3D-427D-BA61-053DB0B94967}
